FIX: Fix incorrect proof in join_cover
This is an API break, because the previous proof was unsound. Like the
bug points out, in join cover we take (self, other) and extend self
until the end of other; if other ends before self, we don't gain
any "cover" from other, and thus the proof is the same as self (making
the method quite useless).

Added traits for checking the proof value in dynamic code, since
completest really doesn't work well right now.